% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dhms}
\alias{parse_dhms}
\title{Parsing dhms values}
\usage{
parse_dhms(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
These functions convert character vectors to objects of the [dhms] class.
`NA` values are supported.
}
\details{
`parse_dhms()` accepts values of the form `"Dd HH:MM:SS"`, with optional day,
fractional seconds and supports negative times.
}
\examples{
time_str = c("12:34:56", "1d 12:34:56", "12:34:56.789", "1d 12:34:56.001",
             "-1d 12:34:56.001", "- 1d 12:34:56.001")
print(parse_dhms(time_str))
}
