% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_dhms}
\alias{round_dhms}
\title{Round or truncate to a multiple of seconds}
\usage{
round_dhms(x, secs)
}
\arguments{
\item{x}{A vector of class [dhms]}

\item{secs}{Multiple of seconds, a positive numeric. Values less than one
are supported}
}
\value{
The input, rounded or truncated to the nearest multiple of `secs`
}
\description{
Convenience functions to round or truncate to a multiple of seconds.
}
\examples{
round_dhms(as.dhms("12:34:56"), 5)
round_dhms(as.dhms("12:34:56"), 60)
round_dhms(as.dhms("1d 12:34:56"), 60)
round_dhms(as.dhms("-1d 12:34:56"), 60)
}
