% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhms.R
\name{dhms}
\alias{dhms}
\alias{dhms}
\alias{as.dhms}
\alias{as.dhms.default}
\alias{as.dhms.difftime}
\alias{as.dhms.numeric}
\alias{as.dhms.charcter}
\alias{as.character.dhms}
\alias{format.dhms}
\alias{print.dhms}
\title{A simple class for storing time-of-day values}
\usage{
dhms(seconds = NULL, minutes = NULL, hours = NULL, days = NULL)

as.dhms(x, ...)

\method{as.dhms}{default}(x, ...)

\method{as.dhms}{difftime}(x, ...)

\method{as.dhms}{numeric}(x, ...)

\method{as.dhms}{charcter}(x, ...)

\method{as.character}{dhms}(x, ...)

\method{format}{dhms}(x, ...)

\method{print}{dhms}(x, ...)
}
\arguments{
\item{seconds, minutes, hours, days}{Time since ZT0. No bounds checking is
performed.}

\item{x}{An object.}

\item{...}{Arguments passed on to further methods.}
}
\description{
The values are stored as a [numeric] vector with a custom class,
and always with "seconds" as unit for robust coercion to numeric.
Supports construction from time values, coercion to and from
various data types, and formatting.  Can be used as a regular column in a
data frame.
}
\details{
For `dhms`, all arguments must have the same length or be
  `NULL`.  Odd combinations (e.g., passing only `seconds` and
  `hours` but not `minutes`) are rejected.
}
\examples{
\dontrun{
#' hms(56, 34, 12)
as.hms(1)
as.hms("12:34:56")
as.hms(Sys.time())
as.POSIXct(hms(1))
  # Will raise an error
  data.frame(a = hms(1))
d <- data.frame(hours = 1:3)
d$hours <- hms(hours = d$hours)
d
}
}
